Fixed a few bugs

+++ src/models/EnergyLevel.java
@@ -3,6 +3,6 @@ package models;
 public class EnergyLevel {
 	
     public int energyThreshold;
-    public int costPerMwh ;
+    public double costPerMwh ;
     public double tonCo2PerMwh;
 }
